+++ R/TextReuseTextDocument.R
@@ -34,7 +34,7 @@
 #'   objects, which can result in significant memory savings for large corpora.
 #'
 #'   If \code{skip_short = TRUE}, this function will return \code{NULL} for very
-#'   short or empty documents. A very short document is one where there are two
+#'   short or empty documents. A very short document is one where there are too
 #'   few words to create at least two n-grams. For example, if five-grams are
 #'   desired, then a document must be at least six words long. If no value of
 #'   \code{n} is provided, then the function assumes a value of \code{n = 3}. A